+++ app/src/main/res/layout/activity_meet_up_creation.xml
@@ -62,12 +62,26 @@
                 android:layout_height="wrap_content"
                 android:text="@string/meetup_creation_meetup_icon" />
 
-            <ImageView
-                android:id="@+id/imageView"
-                android:layout_width="141dp"
-                android:layout_height="wrap_content"
-                app:srcCompat="@drawable/ic_launcher_background"
-                tools:srcCompat="@tools:sample/avatars" />
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:orientation="horizontal">
+
+                <ImageView
+                    android:id="@+id/iv_Icon"
+                    android:layout_width="141dp"
+                    android:layout_height="wrap_content"
+                    app:srcCompat="@drawable/ic_launcher_background"
+                    tools:srcCompat="@tools:sample/avatars" />
+
+                <Button
+                    android:id="@+id/button3"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_weight="1"
+                    android:text="Button" />
+
+            </LinearLayout>
 
             <View
                 android:id="@+id/divider2"
@@ -91,28 +105,51 @@
                 android:id="@+id/button2"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:onClick="onTimeSelectButton"
-                android:text="Button" />
+                android:onClick="onStartTimeSelectButton"
+                android:text="@string/meetup_creation_meetup_start_date_bt" />
 
             <TextView
-                android:id="@+id/textView5"
+                android:id="@+id/textView"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@string/meetup_creation_meetup_duration" />
+                android:text="@string/meetup_creation_meetup_end_date" />
 
-            <SeekBar
-                android:id="@+id/seekBar2"
-                style="@style/Widget.AppCompat.SeekBar.Discrete"
+            <TextView
+                android:id="@+id/EndTime"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:max="10"
-                android:progress="3" />
+                android:text="TextView" />
+
+            <Button
+                android:id="@+id/button"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:onClick="onEndTimeSelectButton"
+                android:text="@string/meetup_creation_meetup_end_date_bt" />
 
             <View
                 android:id="@+id/divider3"
                 android:layout_width="match_parent"
                 android:layout_height="1dp"
                 android:background="?android:attr/listDivider" />
+
+            <TextView
+                android:id="@+id/textView7"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="TextView" />
+
+            <TextView
+                android:id="@+id/textView8"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="TextView" />
+
+            <Button
+                android:id="@+id/button4"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="Button" />
 
         </LinearLayout>
     </ScrollView>
