+++ app/src/main/res/values/strings.xml
@@ -18,9 +18,14 @@
     <string name="meetup_creation_meetup_description">Event Description</string>
     <string name="meetup_creation_meetup_icon">Icon</string>
     <string name="meetup_creation_meetup_start_date">Start Date</string>
+    <string name="meetup_creation_meetup_end_date">End Date</string>
     <string name="meetup_creation_meetup_duration">Duration</string>
     <string name="meetup_creation_meetup_location">Location</string>
     <string name="meetup_creation_meetup_has_participant_limit">Has Participant Limit?</string>
     <string name="meetup_creation_meetup_participant_limit">Participant Limit</string>
     <string name="meetup_creation_meetup_tags">Tags</string>
+    <string name="meetup_creation_meetup_start_date_bt">Select Start Date</string>
+    <string name="meetup_creation_meetup_end_date_bt">Select End Date</string>
+
+    <string name="date_long_format">EEEE, MMMM d, yyyy \'at\' h:mm a</string>
 </resources>
